refactor(web): promote shared ranking and link-path helpers to common/
Both the file-tree Copy Link action (store/) and the Markdown editor's
Link-to-file picker (features/documents/) need Quick Open's fuzzy-ranking
utilities and the note-relative path helpers, but features/documents and
store may not import a sibling feature or reach into a feature respectively.
Move fuzzy.ts/quickOpen.ts and relativeLinkPath/portableImageMarkdownPath to
common/lib/ so every consumer shares one implementation instead of forking
one to satisfy the import boundary. milkdown/paths.ts re-exports the two
functions for its existing in-feature callers.
