+++ web-src/src/common/lib/quickOpen.ts
