+++ web-src/src/features/search/components/ManagedQuickOpen.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { commandDefinitions, rankCommandPalette, routeQuickAccess } from '@/features/search/lib/commandPalette';
-import { rankQuickOpen } from '@/features/search/lib/quickOpen';
+import { rankQuickOpen } from '@/common/lib/quickOpen';
 import { useAppActions, useWorkspace } from '@/store/contexts/AppContext';
 import { openLibrarySearch } from '@/common/lib/librarySearchTrigger';
 import { openSettings } from '@/common/lib/settingsTrigger';
